[terraform] Reuse new variables to define OS image for GCP
The same value for the compute engine OS images was repeated
over an over in different variables when, by default, we only
need it once because we use the same OS image.

To avoid repeating ourselves when the OS image is set for
the compute engines on GCP, we have define a new local
named `machine_image` value that stores the name of the
OS image using the new two variables `os_image_family` (e.g. debian)
and `os_image_version` (e.g. 11).

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>