[terraform] Install Ops Agent on GCP compute engine instances
So far we were using ansible to install on GCP the logging and
monitoring agent (Ops Agent). This agent is specific to GCP,
so it won't work on other cloud solutions. Therefore,
we consider that it's better to install it using terraform,
keeping the provisioning of the platform as generic as possible.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>